// src/v1/auth/types/google-oauth-response.type.ts
export interface GoogleOAuthTokenResponse {
	access_token: string;
	expires_in: number;
	scope: string;
	token_type: string;
	id_token: string;
}

export interface GoogleOAuthUserInfoResponse {
	id: string;
	email: string;
	verified_email: boolean;
	name: string;
	given_name: string;
	family_name: string;
	picture: string;
	locale: string;
}